# [ci-visibility] Add `gatherPayloadsMaxTimeout`, a faster version of `gatherPayloads`