--- a/integration-tests/ci-visibility-intake.js
+++ b/integration-tests/ci-visibility-intake.js
@@ -169,6 +169,38 @@ class FakeCiVisIntake extends FakeAgent {
     infoResponse = DEFAULT_INFO_RESPONSE
   }
 
+  // Similar to gatherPayloads but resolves if enough payloads have been gathered
+  // to make the assertions pass. It times out after maxGatheringTime so it should
+  // always be faster or as fast as gatherPayloads
+  gatherPayloadsMaxTimeout (payloadMatch, onPayload, maxGatheringTime = 15000) {
+    const payloads = []
+    return new Promise((resolve, reject) => {
+      const timeoutId = setTimeout(() => {
+        try {
+          onPayload(payloads)
+          resolve()
+        } catch (e) {
+          reject(e)
+        } finally {
+          this.off('message', messageHandler)
+        }
+      }, maxGatheringTime)
+      const messageHandler = (message) => {
+        if (!payloadMatch || payloadMatch(message)) {
+          payloads.push(message)
+          try {
+            onPayload(payloads)
+            clearTimeout(timeoutId)
+            this.off('message', messageHandler)
+          } catch (e) {
+            // we'll try again when a new payload arrives
+          }
+        }
+      }
+      this.on('message', messageHandler)
+    })
+  }
+
   gatherPayloads (payloadMatch, gatheringTime = 15000) {
     const payloads = []
     return new Promise((resolve, reject) => {