### What does this PR do?
Its API is a bit clumsier (callback vs promise) so I'm going to leave `gatherPayloads` for situations where timing is not that much of an issue. 

### Motivation
Make sure some browser test checks finish as soon as possible, not after the gathering time has passed. 

### Examples

#### Before
* [cypress tests](https://github.com/DataDog/dd-trace-js/actions/runs/4565478581/jobs/8056654326) (2m 36s)
* [playwright tests](https://github.com/DataDog/dd-trace-js/actions/runs/4565478581/jobs/8056654505) (2m 45s)

#### After
* [cypress tests](https://github.com/DataDog/dd-trace-js/actions/runs/4565498010/jobs/8056700268) (2m 18s)
* [playwright tests](https://github.com/DataDog/dd-trace-js/actions/runs/4565498010/jobs/8056700588) (1m 23s)